+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # AutomationLabsSystems Changelog
 
+## v0.1.1
+
+* Add system linearization.
+* Add model evaluation from a system.
+* Improve tests.
 ## v0.1.0
 
 * Initial public release.